add async stuff, but not working. reorganize the config
