+++ fidesctl/fidesctl.toml
@@ -2,5 +2,5 @@
 server_url = "http://fidesctl:8080"
 
 [api]
-database_url = "postgresql+psycopg2://postgres:fidesctl@fidesctl-db:5432/fidesctl"
-test_database_url = "postgresql+psycopg2://postgres:fidesctl@fidesctl-db:5432/fidesctl_test"
+database_url = "postgres:fidesctl@fidesctl-db:5432/fidesctl"
+test_database_url = "postgres:fidesctl@fidesctl-db:5432/fidesctl_test"
