+++ fidesctl/src/fidesapi/crud.py
@@ -12,7 +12,8 @@ from sqlalchemy import update as _update
 from sqlalchemy.dialects.postgresql import Insert as _insert
 from sqlalchemy.exc import SQLAlchemyError
 
-from fidesapi import db_session, errors
+from fidesapi import errors
+from fidesapi.database import database
 from fidesapi.sql_models import SqlAlchemyBase, sql_model_map
 from fideslang import model_map
 
@@ -21,15 +22,17 @@ def get_resource_type(router: APIRouter) -> str:
     "Extracts the name of the resource type from the prefix."
     return router.prefix[1:]
 
-# Update these to use the async sqlalchemy API
+
 # CRUD Functions
-def create_resource(sql_model: SqlAlchemyBase, sql_resource: SqlAlchemyBase) -> Dict:
+async def create_resource(
+    sql_model: SqlAlchemyBase, sql_resource: SqlAlchemyBase
+) -> Dict:
     """Create a resource in the database."""
     with log.contextualize(
         sql_model=sql_model.__name__, fides_key=sql_resource.fides_key
     ):
         try:
-            get_resource(sql_model, sql_resource.fides_key)
+            await get_resource(sql_model, sql_resource.fides_key)
         except errors.NotFoundError:
             log.debug("Resource not found. Inserting.")
         else:
@@ -39,48 +42,48 @@ def create_resource(sql_model: SqlAlchemyBase, sql_resource: SqlAlchemyBase) -> 
             log.bind(error=error.detail["error"]).error("Failed to insert resource")
             raise error
 
-    session = db_session.create_session()
+    session = database.get_db()
 
     with log.contextualize(
         sql_model=sql_model.__name__, fides_key=sql_resource.fides_key
     ):
         try:
             log.debug("Creating resource")
-            session.add(sql_resource)
-            session.commit()
+            await session.add(sql_resource)
+            await session.commit()
         except SQLAlchemyError:
-            session.rollback()
+            await session.rollback()
             error = errors.QueryError()
             log.bind(error=error.detail["error"]).error("Failed to create resource")
             raise error
         finally:
             session.close()
 
-        return get_resource(sql_model, sql_resource.fides_key)
+        return await get_resource(sql_model, sql_resource.fides_key)
 
 
-def get_resource(sql_model: SqlAlchemyBase, fides_key: str) -> Dict:
+async def get_resource(sql_model: SqlAlchemyBase, fides_key: str) -> Dict:
     """
     Get a resource from the databse by its FidesKey.
     """
-    session = db_session.create_session()
+    session = database.get_db()
 
     with log.contextualize(sql_model=sql_model.__name__, fides_key=fides_key):
         try:
             log.debug("Fetching resource")
             sql_resource = (
-                session.query(sql_model)
+                await session.query(sql_model)
                 .filter(sql_model.fides_key == fides_key)
                 .limit(1)
                 .first()
             )
         except SQLAlchemyError:
-            session.rollback()
+            await session.rollback()
             error = errors.QueryError()
             log.bind(error=error.detail["error"]).error("Failed to fetch resource")
             raise error
         finally:
-            session.close()
+            await session.close()
 
         if sql_resource is None:
             error = errors.NotFoundError(sql_model.__name__, fides_key)
@@ -90,60 +93,62 @@ def get_resource(sql_model: SqlAlchemyBase, fides_key: str) -> Dict:
         return sql_resource
 
 
-def list_resource(sql_model: SqlAlchemyBase) -> List:
+async def list_resource(sql_model: SqlAlchemyBase) -> List:
     """
     Get a list of all of the resources of this type from the database.
     """
-    session = db_session.create_session()
+    session = database.get_db()
 
     with log.contextualize(sql_model=sql_model.__name__):
         try:
             log.debug("Fetching resources")
-            sql_resource = session.query(sql_model).all()
+            sql_resource = await session.query(sql_model).all()
         except SQLAlchemyError:
-            session.rollback()
+            await session.rollback()
             error = errors.QueryError()
             log.bind(error=error.detail["error"]).error("Failed to fetch resources")
             raise error
         finally:
-            session.close()
+            await session.close()
 
         return sql_resource
 
 
-def update_resource(
+async def update_resource(
     sql_model: SqlAlchemyBase, resource_dict: Dict, fides_key: str
 ) -> Dict:
     """Update a resource in the database by its fides_key."""
-    session = db_session.create_session()
+    session = database.get_db()
 
     with log.contextualize(sql_model=sql_model.__name__, fides_key=fides_key):
-        get_resource(sql_model, fides_key)
+        await get_resource(sql_model, fides_key)
         try:
             log.debug("Updating resource")
-            session.execute(
+            await session.execute(
                 _update(sql_model)
                 .where(sql_model.fides_key == fides_key)
                 .values(resource_dict)
             )
-            session.commit()
+            await session.commit()
         except SQLAlchemyError:
-            session.rollback()
+            await session.rollback()
             error = errors.QueryError()
             log.bind(error=error.detail["error"]).error("Failed to update resource")
             raise error
         finally:
             session.close()
 
-        return get_resource(sql_model, resource_dict.get("fides_key"))
+        return await get_resource(sql_model, resource_dict.get("fides_key"))
 
 
-def upsert_resources(sql_model: SqlAlchemyBase, resource_dicts: List[Dict]) -> None:
+async def upsert_resources(
+    sql_model: SqlAlchemyBase, resource_dicts: List[Dict]
+) -> None:
     """
     Insert new resources into the database. If a resource already exists,
     update it by it's fides_key.
     """
-    session = db_session.create_session()
+    session = database.get_db()
 
     with log.contextualize(
         sql_model=sql_model.__name__,
@@ -152,15 +157,15 @@ def upsert_resources(sql_model: SqlAlchemyBase, resource_dicts: List[Dict]) -> N
         try:
             log.debug("Upserting resources")
             insert_stmt = _insert(sql_model).values(resource_dicts)
-            session.execute(
+            await session.execute(
                 insert_stmt.on_conflict_do_update(
                     index_elements=["fides_key"],
                     set_=insert_stmt.excluded,
                 )
             )
-            session.commit()
+            await session.commit()
         except SQLAlchemyError:
-            session.rollback()
+            await session.rollback()
             error = errors.QueryError()
             log.bind(error=error.detail["error"]).error("Failed to upsert resources")
             raise error
@@ -168,18 +173,18 @@ def upsert_resources(sql_model: SqlAlchemyBase, resource_dicts: List[Dict]) -> N
             session.close()
 
 
-def delete_resource(sql_model: SqlAlchemyBase, fides_key: str) -> Dict:
+async def delete_resource(sql_model: SqlAlchemyBase, fides_key: str) -> Dict:
     """Delete a resource by its fides_key."""
-    session = db_session.create_session()
+    session = database.get_db()
 
     with log.contextualize(sql_model=sql_model.__name__, fides_key=fides_key):
-        sql_resource = get_resource(sql_model, fides_key)
+        sql_resource = await get_resource(sql_model, fides_key)
         try:
             log.debug("Deleting resource")
-            session.delete(sql_resource)
-            session.commit()
+            await session.delete(sql_resource)
+            await session.commit()
         except SQLAlchemyError:
-            session.rollback()
+            await session.rollback()
             error = errors.QueryError()
             log.bind(error=error.detail["error"]).error("Failed to delete resource")
             raise error
@@ -208,7 +213,7 @@ for resource_type, resource_model in model_map.items():
         """Create a resource."""
         sql_model = sql_model_map[resource_type]
         sql_resource = sql_model(**resource.dict())
-        return create_resource(sql_model, sql_resource)
+        return await create_resource(sql_model, sql_resource)
 
     @router.get("/", response_model=List[resource_model], name="List")
     async def ls(  # pylint: disable=invalid-name
@@ -216,7 +221,7 @@ for resource_type, resource_model in model_map.items():
     ) -> List:
         """Get a list of all of the resources of this type."""
         sql_model = sql_model_map[resource_type]
-        return list_resource(sql_model)
+        return await list_resource(sql_model)
 
     @router.get("/{fides_key}", response_model=resource_model)
     async def get(
@@ -224,7 +229,7 @@ for resource_type, resource_model in model_map.items():
     ) -> Dict:
         """Get a resource by its fides_key."""
         sql_model = sql_model_map[resource_type]
-        return get_resource(sql_model, fides_key)
+        return await get_resource(sql_model, fides_key)
 
     @router.post("/{fides_key}", response_model=resource_model)
     async def update(
@@ -235,7 +240,7 @@ for resource_type, resource_model in model_map.items():
         """Update a resource by its fides_key."""
         sql_model = sql_model_map[resource_type]
         resource_dict = resource.dict()
-        return update_resource(sql_model, resource_dict, fides_key)
+        return await update_resource(sql_model, resource_dict, fides_key)
 
     @router.delete("/{fides_key}")
     async def delete(
@@ -243,7 +248,7 @@ for resource_type, resource_model in model_map.items():
     ) -> Dict:
         """Delete a resource by its fides_key."""
         sql_model = sql_model_map[resource_type]
-        query_result = delete_resource(sql_model, fides_key)
+        query_result = await delete_resource(sql_model, fides_key)
         return {
             "message": "resource deleted",
             "resource": query_result,
