+++ fidesctl/src/fidesapi/database/database.py
@@ -1,19 +1,30 @@
 """
-Contains all of the logic for spinning up/tearing down the database.
+Contains all of the logic related to the database including connections, setup, teardown, etc.
 """
 import os
 
 from alembic import command
 from alembic.config import Config
+from sqlalchemy.ext.asyncio.session import AsyncSession
 from alembic.migration import MigrationContext
 from loguru import logger as log
 from sqlalchemy_utils.functions import create_database, database_exists
 
 from fidesapi.errors import QueryError
+from fidesapi.database.session import async_session
 from fidesapi.sql_models import sql_model_map, SqlAlchemyBase
 from fidesapi.crud import upsert_resources
 from fideslang import DEFAULT_TAXONOMY
 from fidesctl.core.utils import get_db_engine
+
+
+async def get_db() -> AsyncSession:
+    """
+    Yields async sessions and awaits the commit.
+    """
+    async with async_session() as session:
+        yield session
+        await session.commit()
 
 
 def get_alembic_config(database_url: str) -> Config:
